+++ src/index.js
@@ -16,7 +16,7 @@ export { default as ButtonWithOptions } from './ButtonWithOptions';
 export { default as Checkbox } from './Checkbox';
 export { default as ToggleSwitch } from './ToggleSwitch';
 export { default as Modal } from './Modal';
-export { ModalSelectorLayout } from './ModalSelectorLayout';
+export { default as ModalSelectorLayout } from './ModalSelectorLayout';
 export {
   MessageBoxMarketerialLayout,
   MessageBoxMarketerialLayout as MessageBoxLayout1,
@@ -109,4 +109,4 @@ export { default as CardGalleryItem } from './CardGalleryItem';
 export { default as Carousel } from './Carousel';
 export { default as CalendarPanel } from './CalendarPanel';
 export { default as Proportion } from './Proportion';
-export { default as DropdownPopover } from './DropdownPopover';
+export { default as DropdownBase } from './DropdownBase';