+++ CHANGELOG.md
@@ -19,11 +19,13 @@ Types of changes:
 ### Changed
 
 - `<DropdownLayout/>` - move options validation to propTypes [#2717](https://github.com/wix/wix-style-react/pull/2717)
+- `<DropdownPopover/>` - Rename to `<DropdownBase/>` [#2730](https://github.com/wix/wix-style-react/pull/2730)
 
 ### Deprecated
 
 - `<ButtonWithOptions/>`,` <IconWithOptions/>` - deprecated components [#2706](https://github.com/wix/wix-style-react/pull/2706)
 - `<Dropdown/>` - Add support for Controlled `selectedId` [#2719](https://github.com/wix/wix-style-react/pull/2719)
+- `<MultiSelect/>` - New Api [#2752](https://github.com/wix/wix-style-react/pull/2752)
 
 ### Fixed